backend/requirements.txt
injector
autobahn[asyncio]
redis
python-dateutil